refactor pkg inside internal
