+++ internal/cli/command/execution/execution_test.go
@@ -4,7 +4,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"octavius/internal/cli/daemon"
 	"octavius/internal/cli/logger"
-	protobuf "octavius/pkg/protofiles/client_CP"
+	protobuf "octavius/internal/pkg/protofiles/client_CP"
 	"testing"
 )
