+++ internal/cli/daemon/client.go
@@ -6,7 +6,7 @@ import (
 	"io"
 	"octavius/internal/cli/client"
 	"octavius/internal/cli/config"
-	protobuf "octavius/pkg/protofiles/client_CP"
+	protobuf "octavius/internal/pkg/protofiles/client_CP"
 	"time"
 
 	"github.com/gogo/protobuf/jsonpb"
