+++ internal/control_plane/server/metadata/repository/metadata.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"errors"
 	"octavius/internal/control_plane/db/etcd"
-	protobuf "octavius/pkg/protofiles/client_CP"
+	protobuf "octavius/internal/pkg/protofiles/client_CP"
 
 	"github.com/gogo/protobuf/proto"
 )
