+++ internal/control_plane/server/metadata/repository/metadata_mock.go
@@ -2,7 +2,7 @@ package repository
 
 import (
 	"context"
-	protobuf "octavius/pkg/protofiles/client_CP"
+	protobuf "octavius/internal/pkg/protofiles/client_CP"
 
 	"github.com/stretchr/testify/mock"
 )
