+++ internal/pkg/protofiles/client_CP/client_cp_service.proto
@@ -1,8 +1,8 @@
 syntax = "proto3";
-option go_package = "octavius/pkg/protofiles/client_CP";
+option go_package = "octavius/internal/pkg/protofiles/client_CP";
 
-import "pkg/protofiles/client_CP/metadata_messages.proto";
-import "pkg/protofiles/client_CP/job_messages.proto";
+import "internal/pkg/protofiles/client_CP/metadata_messages.proto";
+import "internal/pkg/protofiles/client_CP/job_messages.proto";
 
 service ClientCPServices{
    rpc GetStreamLogs(RequestForStreamLog) returns (stream Log);
