+++ internal/pkg/protofiles/client_CP/client_info.proto
@@ -1,5 +1,5 @@
 syntax = "proto3";
-option go_package = "octavius/pkg/protofiles/client_CP";
+option go_package = "octavius/internal/pkg/protofiles/client_CP";
 
 message ClientInfo {
     string client_email=1;
