+++ internal/pkg/protofiles/client_CP/error_message.proto
@@ -1,5 +1,5 @@
 syntax = "proto3";
-option go_package = "octavius/pkg/protofiles/client_CP";
+option go_package = "octavius/internal/pkg/protofiles/client_CP";
 
 message Error{
     int64 error_code=1;
