+++ internal/pkg/protofiles/client_CP/job_messages.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
-option go_package = "octavius/pkg/protofiles/client_CP";
+option go_package = "octavius/internal/pkg/protofiles/client_CP";
 
-import "pkg/protofiles/client_CP/client_info.proto";
+import "internal/pkg/protofiles/client_CP/client_info.proto";
 
 // GetStreamLogs service
 message RequestForStreamLog {
