+++ internal/pkg/protofiles/client_CP/metadata.proto
@@ -1,5 +1,5 @@
 syntax = "proto3";
-option go_package = "octavius/pkg/protofiles/client_CP";
+option go_package = "octavius/internal/pkg/protofiles/client_CP";
 
 message Secret {
     string name=1 [json_name="name"];
