+++ internal/pkg/protofiles/protofiles.go
@@ -1,0 +1,1 @@
+package protofiles
